# Optimize `find`