@Amanieu 2021-07-15
Looking at the disassembly in more detail, it seems that the iterator was confusing LLVM somehow which caused it to unnecessarily unroll the loop one extra time.