@Zoxc 2021-07-15
I was expecting this to only affect debug builds, but I'll take some free performance.